requirements.txt
pyserial>=3.5
anthropicai>=0.40.0
rich>=13.0.0
python-dotenv>=1.0.0
click>=8.1.0
